Refactor ParquetVariantUtil and TestVariantReaders for improved schema handling
